# Conflicts:
#	hoj-springboot/DataBackup/src/main/java/top/hcode/hoj/config/StartupRunner.java
#	hoj-springboot/DataBackup/src/main/java/top/hcode/hoj/controller/oj/ContestController.java
#	hoj-springboot/DataBackup/src/main/java/top/hcode/hoj/judge/remote/RemoteJudgeDispatcher.java
#	hoj-springboot/DataBackup/src/main/java/top/hcode/hoj/manager/admin/contest/AdminContestManager.java
#	hoj-springboot/DataBackup/src/main/java/top/hcode/hoj/manager/admin/problem/AdminProblemManager.java
#	hoj-springboot/DataBackup/src/main/java/top/hcode/hoj/manager/email/EmailManager.java
#	hoj-springboot/DataBackup/src/main/java/top/hcode/hoj/manager/group/problem/GroupProblemManager.java
#	hoj-springboot/DataBackup/src/main/java/top/hcode/hoj/manager/oj/ContestManager.java
#	hoj-springboot/DataBackup/src/main/java/top/hcode/hoj/pojo/vo/ConfigVO.java
#	hoj-springboot/JudgeServer/src/main/java/top/hcode/hoj/controller/JudgeController.java
#	hoj-springboot/JudgeServer/src/main/java/top/hcode/hoj/judge/task/DefaultJudge.java
#	hoj-springboot/JudgeServer/src/main/java/top/hcode/hoj/judge/task/InteractiveJudge.java
#	hoj-springboot/JudgeServer/src/main/java/top/hcode/hoj/judge/task/SpecialJudge.java
#	hoj-springboot/JudgeServer/src/main/java/top/hcode/hoj/judge/task/TestJudge.java
#	hoj-vue/src/common/constants.js
#	hoj-vue/src/common/logo.js
#	hoj-vue/src/store/contest.js
#	hoj-vue/src/views/oj/contest/outside/ACMScoreBoard.vue
#	hoj-vue/src/views/oj/contest/outside/OIScoreBoard.vue
#	hoj-vue/src/views/oj/contest/outside/scoreBoardMixin.js
#	sqlAndsetting/hoj-update.sql

+++ hoj-springboot/JudgeServer/src/main/java/top/hcode/hoj/judge/AbstractJudge.java
@@ -66,7 +66,7 @@ public abstract class AbstractJudge {
                 .stderr(((JSONObject) judgeResult.get("files")).getStr("stderr"))
                 .time(judgeResult.getLong("time") / 1000000) // ns->ms
                 .memory(judgeResult.getLong("memory") / 1024) // b-->kb
-                .exitCode(judgeResult.getInt("exitStatus"))
+                .exitCode(judgeResult.getLong("exitStatus"))
                 .status(judgeResult.getInt("status"))
                 .originalStatus(judgeResult.getStr("originalStatus"))
                 .build();
@@ -83,7 +83,7 @@ public abstract class AbstractJudge {
                 .stderr(((JSONObject) userJudgeResult.get("files")).getStr("stderr"))
                 .time(userJudgeResult.getLong("time") / 1000000) // ns->ms
                 .memory(userJudgeResult.getLong("memory") / 1024) // b-->kb
-                .exitCode(userJudgeResult.getInt("exitStatus"))
+                .exitCode(userJudgeResult.getLong("exitStatus"))
                 .status(userJudgeResult.getInt("status"))
                 .originalStatus(userJudgeResult.getStr("originalStatus"))
                 .build();
@@ -94,7 +94,7 @@ public abstract class AbstractJudge {
                 .stderr(((JSONObject) interactiveJudgeResult.get("files")).getStr("stderr"))
                 .time(interactiveJudgeResult.getLong("time") / 1000000) // ns->ms
                 .memory(interactiveJudgeResult.getLong("memory") / 1024) // b-->kb
-                .exitCode(interactiveJudgeResult.getInt("exitStatus"))
+                .exitCode(interactiveJudgeResult.getLong("exitStatus"))
                 .status(interactiveJudgeResult.getInt("status"))
                 .originalStatus(interactiveJudgeResult.getStr("originalStatus"))
                 .build();
